+++ src/devices/src/virtio/vsock/unix/muxer.rs
@@ -35,6 +35,7 @@ use std::io::Read;
 use std::os::unix::io::{AsRawFd, RawFd};
 use std::os::unix::net::{UnixListener, UnixStream};
 
+use logger::{Metric, METRICS};
 use utils::epoll::{ControlOperation, Epoll, EpollEvent, EventSet};
 
 use super::super::csm::ConnState;
@@ -298,6 +299,7 @@ impl VsockEpollListener for VsockMuxer {
             }
             Err(e) => {
                 warn!("vsock: failed to consume muxer epoll event: {}", e);
+                METRICS.vsock.muxer_event_fails.inc();
             }
         }
     }
@@ -413,6 +415,7 @@ impl VsockMuxer {
 
             _ => {
                 info!("vsock: unexpected event: fd={:?}, evset={:?}", fd, evset);
+                METRICS.vsock.muxer_event_fails.inc();
             }
         }
     }
@@ -492,6 +495,7 @@ impl VsockMuxer {
                 self.rxq.push(MuxerRx::ConnRx(key));
             }
             self.conn_map.insert(key, conn);
+            METRICS.vsock.conns_added.inc();
             Ok(())
         })
     }
@@ -500,6 +504,7 @@ impl VsockMuxer {
     fn remove_connection(&mut self, key: ConnMapKey) {
         if let Some(conn) = self.conn_map.remove(&key) {
             self.remove_listener(conn.as_raw_fd());
+            METRICS.vsock.conns_removed.inc();
         }
         self.free_local_port(key.local_port);
     }
@@ -509,6 +514,8 @@ impl VsockMuxer {
     /// it an RST packet.
     fn kill_connection(&mut self, key: ConnMapKey) {
         let mut had_rx = false;
+        METRICS.vsock.conns_killed.inc();
+
         self.conn_map.entry(key).and_modify(|conn| {
             had_rx = conn.has_pending_rx();
             conn.kill();
@@ -697,6 +704,7 @@ impl VsockMuxer {
                                 "vsock: error updating epoll listener for (lp={}, pp={}): {:?}",
                                 key.local_port, key.peer_port, err
                             );
+                            METRICS.vsock.muxer_event_fails.inc();
                         });
                 }
             } else {
@@ -715,6 +723,7 @@ impl VsockMuxer {
                         "vsock: error updating epoll listener for (lp={}, pp={}): {:?}",
                         key.local_port, key.peer_port, err
                     );
+                    METRICS.vsock.muxer_event_fails.inc();
                 });
             }
         }
@@ -738,6 +747,7 @@ impl VsockMuxer {
 
         if self.killq.is_empty() && !self.killq.is_synced() {
             self.killq = MuxerKillQ::from_conn_map(&self.conn_map);
+            METRICS.vsock.killq_resync.inc();
             // If we've just re-created the kill queue, we can sweep it again; maybe there's
             // more to kill.
             self.sweep_killq();
@@ -1260,6 +1270,12 @@ mod tests {
         let peer_port_last = peer_port_first + defs::MUXER_KILLQ_SIZE;
         let mut listener = ctx.create_local_listener(local_port);
 
+        // Save metrics relevant for this test.
+        let conns_added = METRICS.vsock.conns_added.count();
+        let conns_killed = METRICS.vsock.conns_killed.count();
+        let conns_removed = METRICS.vsock.conns_removed.count();
+        let killq_resync = METRICS.vsock.killq_resync.count();
+
         for peer_port in peer_port_first..=peer_port_last {
             ctx.init_pkt(local_port, peer_port as u32, uapi::VSOCK_OP_REQUEST);
             ctx.send();
@@ -1300,6 +1316,22 @@ mod tests {
         );
         ctx.send();
 
+        // Check that MUXER_KILLQ_SIZE + 2 connections were added
+        // We count +2, because there are two extra connections being
+        // done outside of the loop.
+        assert_eq!(
+            METRICS.vsock.conns_added.count(),
+            conns_added + defs::MUXER_KILLQ_SIZE + 2
+        );
+        // Check that MUXER_KILLQ_SIZE connections were killed
+        assert_eq!(
+            METRICS.vsock.conns_killed.count(),
+            conns_killed + defs::MUXER_KILLQ_SIZE
+        );
+        // No connections should be removed at this point.
+        assert_eq!(METRICS.vsock.conns_removed.count(), conns_removed);
+
+        assert_eq!(METRICS.vsock.killq_resync.count(), killq_resync + 1);
         // After sweeping the kill queue, it should now be synced (assuming the RX queue is larger
         // than the kill queue, since an RST packet will be queued for each killed connection).
         assert!(ctx.muxer.killq.is_synced());
@@ -1311,6 +1343,12 @@ mod tests {
             assert_eq!(ctx.pkt.op(), uapi::VSOCK_OP_RST);
             assert_eq!(ctx.pkt.src_port(), local_port);
         }
+
+        // The connections should have been removed here.
+        assert_eq!(
+            METRICS.vsock.conns_removed.count(),
+            conns_removed + defs::MUXER_KILLQ_SIZE
+        );
 
         // There should be one more packet in the RX queue: the connection response our request
         // that triggered the kill queue sweep.
@@ -1378,5 +1416,74 @@ mod tests {
         // Since initially the connection had two flags set, now there should
         // not be any pending RX in the muxer.
         assert!(!ctx.muxer.has_pending_rx());
+    }
+
+    #[test]
+    fn test_vsock_basic_metrics() {
+        // Save the metrics values that we need tested.
+        let mut tx_packets_count = METRICS.vsock.tx_packets_count.count();
+        let mut rx_packets_count = METRICS.vsock.rx_packets_count.count();
+
+        let tx_bytes_count = METRICS.vsock.tx_bytes_count.count();
+        let rx_bytes_count = METRICS.vsock.rx_bytes_count.count();
+
+        let conns_added = METRICS.vsock.conns_added.count();
+        let conns_removed = METRICS.vsock.conns_removed.count();
+
+        // Create a basic connection.
+        let mut ctx = MuxerTestContext::new("vsock_basic_metrics");
+        let peer_port = 1025;
+        let (mut stream, local_port) = ctx.local_connect(peer_port);
+
+        // Once the handshake is done, we check that the TX bytes count has
+        // not been increased.
+        assert_eq!(METRICS.vsock.tx_bytes_count.count(), tx_bytes_count);
+
+        // Check that one packet was sent through the handshake.
+        assert_eq!(METRICS.vsock.tx_packets_count.count(), tx_packets_count + 1);
+        tx_packets_count = METRICS.vsock.tx_packets_count.count();
+
+        // Check that one packet was received through the handshake.
+        assert_eq!(METRICS.vsock.rx_packets_count.count(), rx_packets_count + 1);
+        rx_packets_count = METRICS.vsock.rx_packets_count.count();
+
+        // Check that a new connection was added.
+        assert_eq!(METRICS.vsock.conns_added.count(), conns_added + 1);
+
+        // Send some data from guest to host.
+        let data = [1, 2, 3, 4];
+        ctx.init_data_pkt(local_port, peer_port, &data);
+        ctx.send();
+
+        // Check that tx_bytes was incremented.
+        assert_eq!(
+            METRICS.vsock.tx_bytes_count.count(),
+            tx_bytes_count + data.len()
+        );
+
+        // Check that one packet was accounted for.
+        assert_eq!(METRICS.vsock.tx_packets_count.count(), tx_packets_count + 1);
+
+        // Send some data from the host to the guest.
+        let data = [1, 2, 3, 4, 5, 6];
+        stream.write_all(&data).unwrap();
+        ctx.notify_muxer();
+        ctx.recv();
+
+        // Check that a packet was received.
+        assert_eq!(METRICS.vsock.rx_packets_count.count(), rx_packets_count + 1);
+
+        // Check that the 6 bytes have been received.
+        assert_eq!(
+            METRICS.vsock.rx_bytes_count.count(),
+            rx_bytes_count + data.len()
+        );
+
+        // Send a connection reset.
+        ctx.init_pkt(local_port, peer_port, uapi::VSOCK_OP_RST);
+        ctx.send();
+
+        // Check that the connection was removed.
+        assert_eq!(METRICS.vsock.conns_removed.count(), conns_removed + 1);
     }
 }